Rename UniversalFunctions.js to helper.js; Rename consequent variables
