+++ lib/CodeGenerator.js
@@ -4,7 +4,7 @@
  */
 var async = require('async');
 var _ = require('underscore');
-var UniversalFunctions = require('../utils/UniversalFunctions');
+var HELPER = require('../utils/helper');
 
 var Services = require('../Services');
 
@@ -15,7 +15,7 @@ exports.generateUniqueCode = function (noOfDigits,userRole, callback) {
     async.series([
         function (cb) {
             //Push All generated codes in excludeAry
-            if (userRole == UniversalFunctions.CONFIG.APP_CONSTANTS.DATABASE.USER_ROLES.USER){
+            if (userRole == HELPER.CONFIG.APP_CONSTANTS.DATABASE.USER_ROLES.USER){
                 Services.UserService.getAllGeneratedCodes(function (err, dataAry) {
                     if (err){
                         cb(err);
@@ -28,7 +28,7 @@ exports.generateUniqueCode = function (noOfDigits,userRole, callback) {
                 })
             }
             else {
-                cb(UniversalFunctions.CONFIG.APP_CONSTANTS.STATUS_MSG.ERROR.IMP_ERROR)
+                cb(HELPER.CONFIG.APP_CONSTANTS.STATUS_MSG.ERROR.IMP_ERROR)
             }
         }, function (cb) {
             //Generate Random Code of digits specified
